CMake: Properly evaluate option values
In CMake as soon as an option is delcared, a variable for it is defined.
In other words, after a `option(MyOpt "MyOpt explanation" OFF)` a
condition like `if(DEFINED MyOpt)` will always yield TRUE.

The proper way to do conditional actions based on the setting of an
option is simply to use `if(MyOpt)`.