Run Conan with Python 3 virtual environment
This ensures that every time CMake invokes Conan, it runs using
Python 3. This virtual environment can be set up in the same way on
every (I think) Travis CI Mac and Linux host, so we can remove all
Python-specific code from the jobs matrix.

It may be a mistake to move 'pip install conan' from the 'install' phase
to the 'before_script' phase, but I think this will simplify the jobs
matrix in the future. Some jobs may need to override the 'install'
phase, and if the Python3 venv code is overridden, those jobs would need
to duplicate all of that code in order to run properly.