Use venv to set default python to python 3
Why:
When Conan version 2.0 is released, it will require Python 3 to run.
When that happens, this project's CI builds may break when Conan
runs using Python 2.

How:
This leverages `venv` to create a virtual environment that will
always use Python version 3 whenever `python` or `pip` are run.
This solution works the same way on Linux, Mac, and Windows,
and it works no matter what version of Python is the default.

Why remove the `--user` flag:
After venv has been run, all future invocations of `pip` will apply
only to the local virtual environment, and not the system as a
whole, so the `--user` flag is not allowed.

Other possible solutions:
* An alternate and much simpler solution would be to use more recent
images of Ubuntu and MacOS that use Python 3 as the default version
of Python, for instance, Ubuntu 20.04. I am hesitant to recommend
this solution because it breaks compatibility with older versions of
Ubuntu and Mac.
* Another possibility would be `update-alternatives`, which is a
great way to manage different versions of programs on Ubuntu.
To my knowledge, it doesn't exist on Mac.
* Travis appears to use `pyenv` to set the desired version of Python
for Python projects. I have been unable to get `pyenv` to install
the same version of Python on both Linux and Mac without writing a
lot of platform-dependent code.
* Use `virtualenv` instead of `venv`. `virtualenv` is preinstalled
on both Linux and Mac travis-ci images, so this would eliminate the
need to `apt-get install python3-venv` on Linux. This would work,
but `virtualenv` is meant for Python 2 package management. The
Python docs specifically recommend using `venv` for package
management when using Python 3.