Fix Conan invocation on Appveyor
We need to specify the Python3 directory in the PATH to invoke Conan
properly.

I have moved the PATH modification line to before the `pip install`
line. When you run `pip install conan`, you get a warning that says that
the Conan script is not in your path; this fixes the underlying cause,
which means fewer warnings.