Update multiple Conan dependencies.
The `fmt` update is meant to address a bug reported in #81.

Note that `spdlog/1.8.0` is dependent on `fmt/7.0.3`.

Related to #65: specific dependency versions supersede #65, but this
does not supplant the move to a separate file. Unless these
dependencies are moved to a separate file, the process of bumping these
dependencies cannot be easily automated. PRs like this one will be
necessary to keep packages up to date.