Add assertions to fuzz test
Add an example for using assertions within fuzz tests. This might be helpful since there is the caveat that when using assertions during fuzzing, one has to ensure that those assertions are not disabled, e.g., via `NDEBUG`. For larger projects one should factor out this testing for assertions but there are multiple ways to do this so inlining might be a good option for the sake of this exemplary file.