Add build options to Dockerfile
This allows users to specify which versions of Clang and GCC they wish
to install, and specify which compilers to bind to the `CC` and `CXX`
environment variables.

This also reorganizes the Dockerfile for readability, and adds a few
packages that are helpful for building the GUI examples.

This also updates the README.md to reflect changes to the Dockerfile.