Add build_examples.sh script
This script installs several packages that are necessary to make the
GUI examples build successfully.

Every single GUI example depends on a couple of packages being
installed, and for whatever reason, the Conan recipes just don't build
properly if you don't install them yourself. This script adds the
missing dependencies so that everything will build properly in the
Docker container.

It might make more sense to install all of these dependencies from the
Dockerfile, but they take up a lot of disk space and not everyone is
going to need all of them. I tried to keep the packages specified in the
Dockerfile down to the intersection of all six sets of packages, and
then depend on the build_examples.sh script to fill in whatever is
missing. This script should also serve as helpful documentation for
anyone having trouble building specific examples locally.