Add conanfile.py
This moves all hardcoded dependencies out of CMake code and into
conanfile.py. Python code is responsible for conditionally adding
dependencies, based on options passed to CMake.

This puts all the dependencies in one place, and makes the conditional
dependency additions much simpler to understand. According to Conan
documentation, this is the recommended way to handle conditional
dependencies.

Supersedes #124