Replace `sudo` with appropriate env variable
In an earlier commit, I installed `sudo` to allow Conan to install
packages as root. This worked, but it's a misuse of Conan. Conan allows
you to set the environment variable `CONAN_SYSREQUIRES_SUDO=0`; this
tells Conan that it already has root privileges, and it doesn't need to
prepend `sudo` to install commands.

This change is probably unnecessary, but I would prefer not to install
programs that are not absolutely necessary.