Add valid c++ identifier project name variable and use it
This commit replaces every hyphen with a underscore in the project name
to create a safe project name that can be used in C++ identifiers. This
is also used for the CMake project name.

This commit also adds the dot as a forbidden symbol and replaces it
with the underscore for a safe project name.